1. No uppercase letters in file names, this is not our convention and causes problems with cross-platform teams. 2. Moved and cleaned up components to match rangle/typescript-react-redux-starter#31.
Connected to rangle/rangle-starter#23
