+++ src/api/mock/users.js
@@ -1,0 +1,14 @@
+export default [
+  {
+    Username: 'admin',
+    Password: 'superuser',
+  },
+  {
+    Username: 'guest',
+    Password: 'letmein',
+  },
+  {
+    Username: 'user',
+    Password: 'pass',
+  },
+];
