--- src/api/mock/users.json
@@ -1,14 +1,0 @@
-[
-  {
-    "Username": "admin",
-    "Password": "superuser"
-  },
-  {
-    "Username": "guest",
-    "Password": "letmein"
-  },
-  {
-    "Username": "user",
-    "Password": "pass"
-  }
-]
