+++ src/components/button.js
@@ -1,24 +1,20 @@
 import React from 'react';
 
-const Button = (props) => {
-  const {
-    children,
-    className = '',
-    style = {},
-    type = 'button',
-    onClick,
-  } = props;
-
-  return (
-    <button
-      type={ type }
-      className={ `btn btn-primary  ${ className }` }
-      style={{ ...styles.base, ...style }}
-      onClick={ onClick }>
-      { children }
-    </button>
-  );
-};
+const Button = ({
+  children,
+  className = '',
+  style = {},
+  type = 'button',
+  onClick,
+}) => (
+  <button
+    type={ type }
+    className={ `btn btn-primary  ${ className }` }
+    style={{ ...styles.base, ...style }}
+    onClick={ onClick }>
+    { children }
+  </button>
+);
 
 const styles = {
   base: {},
@@ -37,6 +33,5 @@ Button.defaultProps = {
   type: 'button',
   style: {},
 };
-
 
 export default Button;
