+++ src/components/content.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const Content = ({ children, style = {}, isVisible }) => (
+  <div
+    className={ `mt3 p1` }
+    style={{ ...styles.base, style }}>
+    { isVisible ? children : null }
+  </div>
+);
+
+const styles = {
+  base: {},
+};
+
+export default Content;
