+++ src/components/form-group.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const FormGroup = ({ children, style = {}, className = '' }) => (
+  <div className={ `p2 ${ className }` } style={{ ...styles.base, ...style }}>
+    { children }
+  </div>
+);
+
+const styles = {
+  base: {},
+};
+
+export default FormGroup;
