+++ src/components/form-label.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const FormLabel = ({ children, style = {} }) => (
+  <label style={{ ...styles.base, ...style }}>
+    { children }
+  </label>
+);
+
+const styles = {
+  base: {},
+};
+
+export default FormLabel;
