+++ src/components/logo.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import LogoImage from '../assets/rangleio-logo.svg';
+
+const Logo = ({ style = {} }) => (
+  <img
+    style={{ ...styles.base, ...style }}
+    src={ LogoImage }
+    alt="Rangle.io" />
+);
+
+const styles = {
+  base: {
+    width: 128,
+  },
+};
+
+export default Logo;
