--- src/components/ui/FormLabel.js
@@ -1,15 +1,0 @@
-import React from 'react';
-
-const FormLabel = ({ children, style = {} }) => {
-  return (
-    <label style={{ ...styles.base, ...style }}>
-      { children }
-    </label>
-  );
-};
-
-const styles = {
-  base: {},
-};
-
-export default FormLabel;
