+++ src/containers/about-page.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { connect } from 'react-redux';
 
-import Container from '../components/ui/Container';
-import Column from '../components/ui/Column';
+import Container from '../components/container';
+import Column from '../components/column';
 
 function mapStateToProps() {
   return {};
