+++ src/containers/app.js
@@ -4,12 +4,12 @@ import { connect } from 'react-redux';
 import { loginUser, logoutUser } from '../actions/session';
 
 import { Link } from 'react-router';
-import Button from '../components/ui/Button';
-import Content from '../components/ui/Content';
-import LoginModal from '../components/login/LoginModal';
-import Logo from '../components/ui/Logo';
-import Navigator from '../components/navigator/Navigator';
-import NavigatorItem from '../components/navigator/NavigatorItem';
+import Button from '../components/button';
+import Content from '../components/content';
+import LoginModal from '../components/login-modal';
+import Logo from '../components/logo';
+import Navigator from '../components/navigator';
+import NavigatorItem from '../components/navigator-item';
 
 function mapStateToProps(state) {
   return {
