+++ src/utils/is-promise.js
